Added agent states capabilities Massive refactor of bad include files
