+++ platformio.ini
@@ -51,3 +51,8 @@ build_src_filter = -<*> +<examples/velocity_controller/velocity_controller.cpp>
 lib_deps = 
     askuric/Simple FOC@^2.2
     https://github.com/sparkfun/SparkFun_TMAG5273_Arduino_Library
+
+[env:micro_ros_reconnection]
+build_src_filter = -<*> +<examples/micro-ros_reconnection/micro-ros_reconnection.cpp>
+lib_deps = 
+    https://github.com/micro-ROS/micro_ros_platformio
