+++ src/micro_ros__actions.cpp
@@ -1,0 +1,5 @@
+#include "micro_ros__actions.h"
+
+Home_Goal homeGoal;
+Home_Feedback homefeedback;
+Home_Result homeResult;
